replaces prism with shiki as the syntax highlihgting should be done at build-time
