+++ package.json
@@ -51,6 +51,7 @@
     "sanity-plugin-dashboard-widget-document-list": "^2.1.0",
     "sanity-plugin-markdown": "^5.1.0",
     "sanity-plugin-youtube-input": "^0.2.0",
+    "shiki": "^3.2.1",
     "styled-components": "^6.1.16",
     "uuid": "^11.1.0",
     "web-vitals": "^4.2.4"
