+++ src/components/CodeBlock/index.tsx
@@ -1,40 +1,21 @@
-'use client';
-
 import React from 'react';
-import Prism from 'prismjs';
-import 'prismjs/themes/prism-tomorrow.css';
-import 'prismjs/components/prism-typescript';
-import 'prismjs/components/prism-jsx';
-import 'prismjs/components/prism-tsx';
-import 'prismjs/components/prism-css';
-import 'prismjs/components/prism-json';
-import 'prismjs/components/prism-yaml';
-import 'prismjs/components/prism-markup';
-import 'prismjs/components/prism-bash';
-import 'prismjs/components/prism-sql';
+import { highlightCode } from '@/lib/shiki';
+import type { BundledLanguage } from 'shiki';
 
 interface CodeBlockProps {
   code?: string;
-  language?: string;
+  language?: BundledLanguage;
 }
 
-const CodeBlock: React.FC<CodeBlockProps> = ({ code, language }) => {
+export default async function CodeBlock({ code, language }: CodeBlockProps) {
   if (!code) return null;
 
-  const highlightedCode = Prism.highlight(
-    code,
-    Prism.languages[language || 'plaintext'] || Prism.languages.plaintext,
-    language || 'plaintext'
-  );
+  const highlightedCode = await highlightCode(code, language);
 
   return (
-    <pre className="bg-gray-900 text-gray-100 rounded-lg overflow-x-auto my-4 clear-both">
-      <code
-        className={`language-${language || 'plaintext'}`}
-        dangerouslySetInnerHTML={{ __html: highlightedCode }}
-      />
-    </pre>
+    <div
+      className="rounded-lg overflow-x-auto my-4 clear-both"
+      dangerouslySetInnerHTML={{ __html: highlightedCode }}
+    />
   );
-};
-
-export default CodeBlock;
+}
